+++ packages/proof-skin/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @refrakt-md/proof-skin
 
+## 0.36.0
+
+### Patch Changes
+
+- @refrakt-md/skeleton@0.36.0
+
 ## 0.35.0
 
 ### Patch Changes